requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
cryptography==46.0.1
deprecation==2.1.0
Flask==3.1.1
flask-cors==6.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
postgrest==2.20.0
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
realtime==2.20.0
sniffio==1.3.1
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
typing-inspection==0.4.1
typing_extensions==4.14.0
websockets==15.0.1
Werkzeug==3.1.3

python-dotenv

